requirements.txt
opencv-python
face_recognition
numpy
mysql-connector-python
cryptography